@@ -1,72 +1,34 @@
 {
-  "name": "react-sample-app",
-  "version": "0.1.0",
+  "name": "create_quiz",
   "private": true,
-  "dependencies": {
-    "@google/generative-ai": "^0.21.0",
-    "@mantine/carousel": "^7.16.2",
-    "@mantine/charts": "^7.16.2",
-    "@mantine/code-highlight": "^7.16.2",
-    "@mantine/core": "^7.16.2",
-    "@mantine/dates": "^7.16.2",
-    "@mantine/dropzone": "^7.16.2",
-    "@mantine/form": "^7.16.2",
-    "@mantine/hooks": "^7.16.2",
-    "@mantine/modals": "^7.16.2",
-    "@mantine/notifications": "^7.16.2",
-    "@mantine/nprogress": "^7.16.2",
-    "@mantine/spotlight": "^7.16.2",
-    "@mantine/tiptap": "^7.16.2",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@tiptap/extension-link": "^2.11.5",
-    "@tiptap/pm": "^2.11.5",
-    "@tiptap/react": "^2.11.5",
-    "@tiptap/starter-kit": "^2.11.5",
-    "cors": "^2.8.5",
-    "dayjs": "^1.11.13",
-    "dotenv": "^16.4.5",
-    "embla-carousel-react": "^7.1.0",
-    "express": "^4.21.2",
-    "mysql2": "^3.12.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-scripts": "5.0.1",
-    "recharts": "^2.15.1",
-    "uuid": "^11.0.5",
-    "web-vitals": "^2.1.4"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "dev": "nodemon app.jsx",
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview",
     "server": "node ./backend/restful.js"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "@google/generative-ai": "^0.24.0",
+    "@tailwindcss/vite": "^4.0.15",
+    "cors": "^2.8.5",
+    "express": "^4.21.2",
+    "mysql2": "^3.14.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "tailwindcss": "^4.0.15"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "postcss": "^8.5.1",
-    "postcss-preset-mantine": "^1.17.0",
-    "postcss-simple-vars": "^7.0.1"
+    "@eslint/js": "^9.21.0",
+    "@types/react": "^19.0.10",
+    "@types/react-dom": "^19.0.4",
+    "@vitejs/plugin-react": "^4.3.4",
+    "eslint": "^9.21.0",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "globals": "^15.15.0",
+    "vite": "^6.2.0"
   }
 }
