@@ -1,10 +1,8 @@
-import styles from './css/Question.module.css';
-
 export default function QuestionDisplay({ quizIndex, titleData }) {
   return (
-    <div className={styles.title}>
-      <h2>問題{quizIndex + 1}</h2>
-      <p>{titleData}</p>
+    <div>
+      <h2 className="text-center text-2xl font-bold mb-5">問題{quizIndex + 1}</h2>
+      <p className="m-5 text-xl">{titleData}</p>
     </div>
   );
 }
