@@ -1,9 +1,7 @@
 import { useState } from 'react';
-
 import QuestionDisplay from './QuestionDisplay.jsx';
 import SelectDisplay from './SelectDisplay.jsx';
 import AnswerDisplay from './AnswerDisplay.jsx';
-import styles from './css/Quiz.module.css';
 
 export default function Quiz({ quizIndex, questionData }) {
   const [checkedValue, setCheckedValue] = useState(null);
@@ -13,7 +11,7 @@ export default function Quiz({ quizIndex, questionData }) {
   };
 
   return (
-    <div className={`${styles.quiz}`} id="quiz">
+    <div id="quiz">
       {/* 問題を表示 */}
       <QuestionDisplay quizIndex={quizIndex} titleData={questionData.question} />
       {/* 選択肢を表示 */}
