@@ -1,8 +1,6 @@
 import { useContext } from 'react';
 import { resultContext } from './QuizContext.jsx';
 
-import styles from './css/Result.module.css';
-
 export default function QuizResult() {
   const { result } = useContext(resultContext);
   const totalQuestions = result.length;
@@ -16,17 +14,17 @@ export default function QuizResult() {
 
   return (
     <>
-      <div className={styles.result}>
-        <h1 className={styles.title}>リザルト</h1>
-        <p className={styles.detail}>{totalQuestions}問中{totalCorrect}正解</p>
+      <div className="m-10 text-center">
+        <h2 className="text-3xl font-bold">リザルト</h2>
+        <p className="mb-5 text-xl">{totalQuestions}問中{totalCorrect}正解</p>
         {result.map((questionResult, index) => {
           return (
-            <div key={index} className={styles.question}>
-              <p className={styles.questionNumber}>{index + 1}問目</p>
-              <p className={styles.questionTitle}>{questionResult.question}</p>
-              <div className={styles.correct}>
+            <div key={index} className="my-5 pb-3 border-b border-b-stone-300">
+              <p className="text-xl font-bold">{index + 1}問目</p>
+              <p className="text-lg mb-2">{questionResult.question}</p>
+              <div>
                 <p>正解 : {questionResult.answer}</p>
-                <p className={questionResult.correct ? styles.correctSelect : styles.incorrectSelect}>あなたの答え : {questionResult.select}</p>
+                <p>あなたの答え : {questionResult.select}</p>
               </div>
             </div>
           )
